requirements.txt
pip
setuptools
wheel
Kekik

PyYAML
PyAudio
SpeechRecognition
beepy
gTTS
playsound
pydub
ffmpeg-python
openai

Pyrogram
TgCrypto

httpx
tls-client